Criação da tela de Artistas

+++ lib/service/artista.dart
@@ -1,12 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:ost_tracker_og/models/artista_model.dart';
+import 'package:ost_tracker_og/service/genero.dart';
 import 'package:ost_tracker_og/service/usuario.dart';
 
 class ArtistaFirestore {
   final FirebaseAuth _auth = FirebaseAuth.instance;
   final FirebaseFirestore _db = FirebaseFirestore.instance;
   final UsuarioFirestore _user = UsuarioFirestore();
+  final GeneroFirestore _genero = GeneroFirestore();
 
   Future<List<Artista>> getArtistas() async {
     List<Artista> artistas = [];
@@ -17,6 +19,10 @@ class ArtistaFirestore {
     for (var artista in res.docs) {
       Map<String, dynamic> map = artista.data();
       map["id"] = artista.id;
+      map["generos"] = Future.wait(List.from(map["generos"])
+          .map((e) async => await _genero.getOnlyGenero(e))
+          .toList());
+      print(map["generos"]);
       artistas.add(Artista.fromMap(map));
     }
     return artistas;
@@ -32,17 +38,19 @@ class ArtistaFirestore {
   }
 
   editArtista(Artista artista) async {
+    if (artista.dono != _auth.currentUser!.uid) return;
     await _db
         .doc("/artistas/${artista.id}")
         .update(Artista.toMap(artista))
         .onError((error, stackTrace) => throw Exception(error));
   }
 
-  deletArtista(String id) async {
+  deleteArtista(Artista artista) async {
+    if (artista.dono != _auth.currentUser!.uid) return;
     await _db
-        .doc("/artistas/$id")
+        .doc("/artistas/${artista.id}")
         .delete()
         .onError((error, stackTrace) => throw Exception(error))
-        .then((value) => _user.deleteArtista(id));
+        .then((value) => _user.deleteArtista(artista.id));
   }
 }
